debug

+++ agedb/uncertain_regression.py
@@ -184,7 +184,7 @@ def train_epoch_uncertain(model, train_loader, val_loader, train_labels, opt, ar
             try:
                 varianc = var_tensor.index_select(0, index=y.squeeze(-1).to(torch.int32))
             except:
-                print(y.squeeze(-1))
+                print(y.squeeze(-1).to(torch.int32))
                 not_in = []
                 for i in y.squeeze(-1).to(torch.int32):
                     if i.item() not in var_list:
